# speed up s2s ~2x