--- a/parlai/agents/seq2seq/seq2seq.py
+++ b/parlai/agents/seq2seq/seq2seq.py
@@ -168,7 +168,7 @@ class Seq2seqAgent(Agent):
 
         # all instances may need some params
         self.truncate = opt['truncate'] if opt['truncate'] > 0 else None
-        self.metrics = {'loss': 0.0, 'num_tokens': 0, 'total_skipped_batches': 0}
+        self.metrics = {'loss': 0.0, 'num_tokens': 0, 'correct_tokens': 0, 'total_skipped_batches': 0}
         self.history = {}
         self.report_freq = opt.get('report_freq', 0.001)
         self.use_person_tokens = opt.get('person_tokens', False)
@@ -409,6 +409,7 @@ class Seq2seqAgent(Agent):
         """Reset metrics for reporting loss and perplexity."""
         self.metrics['loss'] = 0.0
         self.metrics['num_tokens'] = 0
+        self.metrics['correct_tokens'] = 0
 
     def report(self):
         """Report loss and perplexity from model's perspective.
@@ -418,6 +419,7 @@ class Seq2seqAgent(Agent):
         """
         m = {}
         if self.metrics['num_tokens'] > 0:
+            m['token_acc'] = self.metrics['correct_tokens'] / self.metrics['num_tokens']
             m['loss'] = self.metrics['loss'] / self.metrics['num_tokens']
             try:
                 m['ppl'] = math.exp(m['loss'])
@@ -488,11 +490,15 @@ class Seq2seqAgent(Agent):
             try:
                 out = self.model(xs, ys, rank_during_training=cands is not None)
                 # generated response
-                predictions, scores, cand_preds = out[0], out[1], out[2]
+                _preds, scores, cand_preds = out[0], out[1], out[2]
+
                 score_view = scores.view(-1, scores.size(-1))
                 loss = self.criterion(score_view, ys.view(-1))
                 # save loss to metrics
-                target_tokens = ys.ne(self.NULL_IDX).long().sum().item()
+                y_ne = ys.ne(self.NULL_IDX)
+                target_tokens = y_ne.long().sum().item()
+                correct = ((ys == _preds) * y_ne).sum().item()
+                self.metrics['correct_tokens'] += correct
                 self.metrics['loss'] += loss.item()
                 self.metrics['num_tokens'] += target_tokens
                 loss /= target_tokens  # average loss per token