seq2seq was returning its predictions as texts to the teacher to get accuracy and f1 metrics, this was hurting it a surprising amount.

this diff removes that, losing access to accuracy and f1 during training logs, but seq2seq now returns token-accuracy as a useful proxy metric instead.

---
babi:task10k:1 experiments

seq2seq command:
`py examples/train_model.py -t babi:task10k:1 --dict-file /tmp/dict_babi:task10k:1 -bs 128 -eps 20 -dt train:ordered -m seq2seq -hs 128 -esz 128 -nl 2`

fairseq command:
`py examples/train_model.py -t babi:task10k:1 --dict-file /tmp/dict_babi:task10k:1 -bs 128 -eps 20 -dt train:ordered -m fairseq --arch lstm --encoder_embed_dim 128 --encoder-hidden-size 128 --encoder-layers 2 --decoder-embed-dim 128 --decoder-hidden-size 128 --decoder-layers 2 --decoder-out-embed-dim 128 --decoder-attention 0
`

seq2seq before: 77s, 78s
seq2seq after: 40s, 40s
fairseq2seq: 62s, 66s

---
convai2 experiments

seq2seq command:
`py examples/train_model.py -t convai2 --dict-file /tmp/dict_convai2 -bs 128 -m seq2seq -eps 1 -dt train:ordered`

fairseq command:
`py examples/train_model.py -t convai2 --dict-file /tmp/dict_convai2 -bs 128 -eps 1 -dt train:ordered -m fairseq --arch lstm --encoder_embed_dim 128 --encoder-hidden-size 128 --encoder-layers 2 --decoder-embed-dim 128 --decoder-hidden-size 128 --decoder-layers 2 --decoder-out-embed-dim 128 --decoder-attention 0`

seq2seq before: 182s, 174s
seq2seq after: 84s, 82s
fairseq: 125s, 125s