@jaseweston 2018-06-29
maybe it should be a param?

On Fri, Jun 29, 2018, 1:18 PM Stephen Roller <notifications@github.com>
wrote:

> *@stephenroller* approved this pull request.
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/facebookresearch/ParlAI/pull/935#pullrequestreview-133309248>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AKjk-BaK62oV0Qg2NJDa6AKwxcWoo6nJks5uBmFOgaJpZM4U9NxE>
> .
>