@alexholdenmiller 2018-06-29
@jaseweston I think better if not, "train 2x slower to show acc and f1 during training" doesn't seem worth it (you can always use -dt train:evalmode to get those metrics on full train set)